+++ packages/common/src/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from './useBooleanOnce'
 export * from './useFeatureFlag'
 export * from './useRemoteVar'
-export * from './useDownload'
+export * from './useStems'
 export * from './useImageSize'
 export * from './useInstanceVar'
 export * from './useLocalStorage'
